requirements.txt
# Core project dependencies
# Add libraries here as your project grows

# Example common packages (uncomment if needed)
numpy<2
pandas
python-dotenv

scikit-learn
xgboost
shap
statsmodels
shap

matplotlib
seaborn